requirements.txt
Flask==1.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
idna==3.1
itsdangerous==1.1.0
numpy==1.19.5
pandas==1.2.0
pip==20.3.3
pyTigerGraph==0.0.9.6.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
setuptools==51.3.3
six==1.15.0
urllib3==1.26.2
validators==0.18.2
